+++ opengate/geometry/helpers_geometry.py
@@ -8,6 +8,8 @@ from .PolyhedraVolume import *
 from .TrdVolume import *
 from .BooleanVolume import *
 from .RepeatParametrisedVolume import *
+from anytree import RenderTree
+from anytree import Node
 
 volume_type_names = {
     BoxVolume,
@@ -98,3 +100,70 @@ elements_name_symbol = {
     "Silver": "Ag",
     "Tin": "Sn",
 }
+
+
+def pretty_print_tree(tree, geometry):
+    """Print tree"""
+    s = ""
+    for pre, fill, node in RenderTree(tree[gate.__world_name__]):
+        v = geometry[node.name]
+        s += f"{pre}{node.name} {v.type_name} {v.material}\n"
+
+    # remove last break line
+    return s[:-1]
+
+
+def build_tree(user_info_volumes):
+    # world is needed as the root
+    uiv = user_info_volumes
+    if gate.__world_name__ not in uiv:
+        s = f"No world in geometry = {uiv}"
+        gate.fatal(s)
+
+    # build the root tree (needed)
+    tree = {gate.__world_name__: Node(gate.__world_name__)}
+    already_done = {gate.__world_name__: True}
+
+    # build the tree
+    for vol in uiv.values():
+        if vol.name in already_done:
+            continue
+        add_volume_to_tree(uiv, already_done, tree, vol)
+
+    return tree
+
+
+def add_volume_to_tree(user_info_volumes, already_done, tree, vol):
+    # check if mother volume exists
+    uiv = user_info_volumes
+    if vol.mother not in uiv:
+        gate.fatal(
+            f"Cannot find a mother volume named '{vol.mother}', for the volume {vol}"
+        )
+
+    already_done[vol.name] = "in_progress"
+    m = uiv[vol.mother]
+
+    # check for the cycle
+    if m.name not in already_done:
+        add_volume_to_tree(uiv, already_done, tree, m)
+    else:
+        if already_done[m.name] == "in_progress":
+            s = f"Error while building the tree, there is a cycle ? "
+            s += f"\n volume is {vol}"
+            s += f"\n parent is {m}"
+            gate.fatal(s)
+
+    # get the mother branch
+    p = tree[m.name]
+
+    # check not already exist
+    if vol.name in tree:
+        s = f"Node already exist in tree {vol.name} -> {tree}"
+        s = s + f"\n Probably two volumes with the same name ?"
+        gate.fatal(s)
+
+    # create the node
+    n = Node(vol.name, parent=p)
+    tree[vol.name] = n
+    already_done[vol.name] = True
